--- Backend/hello.h
@@ -1,1 +1,0 @@
-void hello();
