+++ Backend/module.py
@@ -2,7 +2,7 @@
 from distutils.core import setup, Extension
 
 VectorModule = Extension('VectorModule',
-                    sources = ['vector.cpp'])
+                    sources = ['main.cpp', 'vectorclass.cpp', 'Matrix.cpp'])
 
 setup (name = 'VectorModule',
        version = '1.0',
